SCRUM-98 feat: Moving logout button from header to sidebar.
