+++ src/widgets/Header/components/index.ts
@@ -1,4 +1,3 @@
 export * from './AuthBtn'
 export * from './LanguageSelect'
-export * from './LogoutModal'
 export * from './NotificationButton'
