+++ src/widgets/Header/hooks/index.ts
@@ -1,2 +1,1 @@
 export { useHomeLink } from './useHomeLink'
-export { useLogoutHandler } from './useLogoutHandler'
