+++ src/widgets/Sidebar/components/LogoutModal/index.tsx
